added ExecutionRootPathResolver interface
required for testing later on
